backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.1
python-multipart==0.0.6
loguru==0.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
itsdangerous==2.1.2

# AI Providers
anthropic==0.39.0
openai==1.54.5

# Google Docs integration
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Google Cloud Firestore for RAG
google-cloud-firestore==2.11.1

# Optional: Redis for Celery (if needed later)
redis==5.0.1
celery==5.3.4
